fix: Compile with GCC14.2

+++ src/simu/linux/simu.c
@@ -5,7 +5,7 @@
 //#define VMICROCHIP
 
 typedef unsigned char uchar;
-typedef unsigned long ulong;
+typedef unsigned long uint32_t;
 
 #define vub unsigned char
 #define vsb signed char
